+++ smart-g60/external-system-communication/src/main/java/com/pingok/external/domain/primary/TblStationLLInfo.java
@@ -5,6 +5,7 @@ import lombok.Data;
 import javax.persistence.Id;
 import javax.persistence.Table;
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 @Data
 @Table(name = "TBL_STATION_LL_INFO")
@@ -21,10 +22,10 @@ public class TblStationLLInfo implements Serializable {
     private String nodeId;
 
     /** 纬度 */
-    private Long latitude;
+    private BigDecimal latitude;
 
     /** 经度 */
-    private Long longitude;
+    private BigDecimal longitude;
 
     /** 节点类型 1 高速公路 2 非现场 */
     private Integer nodeType;
